requirements.txt
backtrader
yfinance
pandas
ta
matplotlib
alpaca-py 
websocket-client
alpaca-trade-api 
python-dotenv
streamlit

